requirements.txt
playwright
fake-useragent
playwright-stealth
pymongo
pydantic
streamlit
apscheduler
requests
beautifulsoup4
tenacity
loguru
playwright-stealth
slack_sdk
tenacity
mecab-python3
retry
pandas
matplotlib
seaborn
scikit-learn